Align WarpKriging/MLPKriging API naming with Kriging convention
Rename parameters to match the basic trio (Kriging/NuggetKriging/NoiseKriging):
- predict: x_new→X_n, withStd→return_stdev, withCov→return_cov, withDeriv→return_deriv
- simulate: uint64_t seed→int seed, x_new→X_n
- update: y_new→y_u, X_new→X_u
- logLikelihoodFun: theta_gp→theta, withGrad→return_grad, withHess→return_hess

+++ bindings/Julia/jlibkriging/csrc/libkriging_c.cpp
@@ -1812,7 +1812,7 @@ int lk_warp_kriging_predict(void* ptr,
 int lk_warp_kriging_simulate(void* ptr, int nsim, int seed, const double* X_n, int m, int d, double* sim_out) {
   try {
     arma::mat X_mat(const_cast<double*>(X_n), m, d, false, true);
-    auto result = static_cast<WarpKriging*>(ptr)->simulate(nsim, static_cast<uint64_t>(seed), X_mat);
+    auto result = static_cast<WarpKriging*>(ptr)->simulate(nsim, seed, X_mat);
     if (sim_out)
       std::memcpy(sim_out, result.memptr(), result.n_elem * sizeof(double));
     return 0;
@@ -2088,7 +2088,7 @@ int lk_mlp_kriging_predict(void* ptr,
 int lk_mlp_kriging_simulate(void* ptr, int nsim, int seed, const double* X_n, int m, int d, double* sim_out) {
   try {
     arma::mat X_mat(const_cast<double*>(X_n), m, d, false, true);
-    auto result = static_cast<MLPKriging*>(ptr)->simulate(nsim, static_cast<uint64_t>(seed), X_mat);
+    auto result = static_cast<MLPKriging*>(ptr)->simulate(nsim, seed, X_mat);
     if (sim_out)
       std::memcpy(sim_out, result.memptr(), result.n_elem * sizeof(double));
     return 0;
