+++ bindings/Python/pylibkriging/src/_pylibkriging/WarpKriging_binding.cpp
@@ -73,7 +73,7 @@ PyWarpKriging::predict(const py::array_t<double>& X_n, bool return_stdev, bool r
 py::array_t<double> PyWarpKriging::simulate(const int nsim, const int seed, const py::array_t<double>& X_n,
                                             const bool will_update) {
   arma::mat mat_X = carma::arr_to_mat_view<double>(X_n);
-  auto result = m_internal->simulate(nsim, static_cast<uint64_t>(seed), mat_X, will_update);
+  auto result = m_internal->simulate(nsim, seed, mat_X, will_update);
   return carma::mat_to_arr(result, true);
 }
